+++ CHANGELOG.rst
@@ -9,6 +9,11 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 Unreleased_
 -----------
 
+Added
+~~~~~
+
+* ``readinto`` method to directly read into a bytearray
+
 Fixed
 ~~~~~